fix: use model dtype for PCG adapter_mask and scaling tensors
The adapter_mask was float32 but model weights are float16, causing
dtype mismatch in torch.mm. Now init with model dtype.

+++ python/sglang/srt/model_executor/piecewise_cuda_graph_runner.py
@@ -279,7 +279,7 @@ class PiecewiseCudaGraphRunner:
         # Pre-allocate LoRA PCG batch info at fixed GPU addresses
         if self.model_runner.server_args.enable_lora:
             self.model_runner.lora_manager.lora_backend.init_pcg_batch_info(
-                self.max_num_tokens
+                self.max_num_tokens, dtype=self.model_runner.dtype,
             )
 
         with enable_piecewise_cuda_graph():
